app/requirements.txt
redis==3.5.3
fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==0.19.0
